requirements.txt

Flask
Flask-CORS
mysql-connector-python
